update theme, add Kalam font
This font is now set to all headings. It does look childish and
nonserious, and perhaps that's what I'm going for. It it harder to read
compare to rest of the site, but I wanted to set it for the marathi
text, and this was the collateral damage so oh well.